src/pyproject.toml
[project]
name = "pyaction"
version = "0.0.1"
description = ""
authors = [{name = "Pierre Nicolas Durette", email = "pndurette@gmail.com"}]
requires-python = ">=3.7"

dependencies = [
    "PyYAML==6.0",
    "Markdown==3.4.1"
]

[project.optional-dependencies]
tests = [
    "pytest ~= 7.2",
]

[tool.coverage.run]
omit = [
    "gtts/tests/*",
    "gtts/tokenizer/tests/*",
]